\name{bgd-package}
\alias{bgd-package}
\alias{bgd}
\docType{package}
\title{
\packageTitle{bgd}
}
\description{
\packageDescription{bgd}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bgd}
\packageIndices{bgd}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bgd}

Maintainer: \packageMaintainer{bgd}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
